[feat] : input 박스 변경 - input 박스의 outline, box-shadow 제거 - search 내용을 미들웨어로부터 받아오도록 변경
