+++ src/app/globals.css
@@ -286,3 +286,8 @@ h6 {
 *::-webkit-scrollbar-thumb:hover {
   background: #a0a5b1;
 }
+
+input {
+  outline: none;
+  box-shadow: none;
+}
